+++ src/test/kotlin/watch/craft/scrapers/ThornbridgeScraperTest.kt
@@ -14,7 +14,7 @@ class ThornbridgeScraperTest {
 
   @Test
   fun `finds all the beers`() {
-    assertEquals(14, ITEMS.size)
+    assertEquals(15, ITEMS.size)
   }
 
   @Test
